data-tooling/src/cli.py
# CLI for Data Dirtying Tool

import click

# Import necessary modules
# from src.bigquery_data import some_function
# from src.csv_utils import another_function

@click.group()
def cli():
    """Main entry point for the CLI"""
    pass

# Placeholder for the data dirtying command
@cli.command()
def dirty_data():
    """Command to dirty a dataset"""
    # Logic will go here
    pass

# Placeholder for the setup command
@cli.command()
def setup():
    """Command to set up the environment"""
    # Logic will go here
    pass

if __name__ == '__main__':
    cli()


