data-tooling/tests/test_cli.py
# Unit tests for the CLI

import pytest
from click.testing import CliRunner
from src.cli import cli

@pytest.fixture
def runner():
    """Fixture providing a Click runner for testing the CLI"""
    return CliRunner()

def test_cli_runner(runner):
    """Basic test to ensure the CLI runs"""
    result = runner.invoke(cli)
    assert result.exit_code == 0

# Placeholder for testing dirty_data command
def test_dirty_data_command(runner):
    """Test for the dirty_data command"""
    result = runner.invoke(cli, ['dirty_data'])
    assert result.exit_code == 0

# Placeholder for testing setup command
def test_setup_command(runner):
    """Test for the setup command"""
    result = runner.invoke(cli, ['setup'])
    assert result.exit_code == 0


